Implemented permissions and authentication for api

+++ expense_tracker/tracker/api_views.py
@@ -1,31 +1,40 @@
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from .models import Expense
 from .serializers import ExpenseSerializer
 
 
 @api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def expense_list(request):
 
     if request.method == 'GET':
-        expenses = Expense.objects.all()
+        expenses = Expense.objects.filter(user=request.user)
         serializer = ExpenseSerializer(expenses, many=True)
         return Response(serializer.data)
 
     elif request.method == 'POST':
         serializer = ExpenseSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            if serializer.validated_data['user'] == request.user:
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            else:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def expense_detail(request, pk):
 
     try:
-        expense = Expense.object.get(pk=pk)
+        expense = Expense.objects.get(pk=pk)
     except Expense.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -34,12 +43,19 @@ def expense_detail(request, pk):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
+        print(expense.user)
         serializer = ExpenseSerializer(expense, data=request.data)
         if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
+            if expense.user == request.user and serializer.validated_data['user'] == request.user:
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
-        expense.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if expense.user == request.user:
+            expense.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
